requirements.txt
openpyxl==3.0.5
Pillow==7.2.0
tqdm==4.49.0
py2app==0.22
lxml==4.5.2
fbs==0.9.0
PyQt5==5.15.1

